requirements.txt
requests==2.32.3
rich==13.7.1
dnspython==2.6.1
